Fix implicit conversion in Str::rrand

+++ src/str.cls.php
@@ -20,7 +20,7 @@ class Str {
 	 * @return string
 	 */
 	public static function rrand( $len, $type = 7 ) {
-		mt_srand( ( double ) microtime() * 1000000 );
+		mt_srand( (int) ( ( double ) microtime() * 1000000 ) );
 
 		switch( $type ) {
 			case 0 :
